server/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg[binary]==3.3.3
PyJWT==2.10.1
pytest==8.3.5